rename result to submission
